+++ module/move/assistant/src/print.rs
@@ -3,137 +3,7 @@ use super::*;
 use core::fmt;
 use former::Former;
 
-// ==
-
-/// A trait for iterating over all rows of a table.
-pub trait TableSize< 'a >
-{
-  /// Returns size of a table.
-  fn table_size( &'a self ) -> [ usize ; 2 ];
-}
-
-/// A trait for iterating over all rows of a table.
-pub trait TableRows< 'a, RowKey, Row, CellKey, Cell >
-where
-  Row : Clone + for< 'cell > Cells< 'cell, CellKey, Cell > + 'a,
-  Cell : fmt::Debug + Clone + 'a,
-{
-  /// Returns an iterator over all rows of the table.
-  fn rows( &'a self ) -> impl IteratorTrait< Item = Row >;
-}
-
-/// Trait returning headers of a table if any.
-pub trait TableHeader< 'a, CellKey, Title >
-where
-  Title : fmt::Debug,
-{
-  /// Returns an iterator over all fields of the specified type within the entity.
-  fn header( &'a self ) -> Option< impl IteratorTrait< Item = ( CellKey, Title ) > >;
-}
-
-/// A trait for iterating over all cells of a row.
-pub trait Cells< 'a, CellKey, Cell >
-where
-  Cell : fmt::Debug + Clone + 'a,
-{
-  /// Returns an iterator over all cells of the row.
-  fn cells( &'a self ) -> impl IteratorTrait< Item = ( CellKey, Cell ) >
-  ;
-}
-
-// ==
-
-impl< 'a, T, RowKey, Row, CellKey, Cell, Title > TableSize< 'a >
-for AsTable< 'a, T, RowKey, Row, CellKey, Cell, Title >
-where
-  Self : TableRows< 'a, RowKey, Row, CellKey, Cell >,
-  Row : Clone + for< 'cell > Cells< 'cell, CellKey, Cell > + 'a,
-  Title : fmt::Debug,
-  Cell : fmt::Debug + Clone + 'a,
-  CellKey : fmt::Debug + Clone,
-{
-  fn table_size( &'a self ) -> [ usize ; 2 ]
-  {
-    let mut rows = self.rows();
-    let nrows = rows.len();
-    let row = rows.next();
-    if let Some( row ) = row
-    {
-      let ncells = row.cells().len();
-      [ nrows, ncells ]
-    }
-    else
-    {
-      [ 0, 0 ]
-    }
-  }
-}
-
-impl< 'a, T, RowKey, Row, CellKey, Cell, Title > TableRows< 'a, RowKey, Row, CellKey, Cell >
-for AsTable< 'a, T, RowKey, Row, CellKey, Cell, Title >
-where
-  T : Fields< 'a, RowKey, Row >,
-  Row : Clone + for< 'cell > Cells< 'cell, CellKey, Cell > + 'a,
-  Title : fmt::Debug,
-  Cell : fmt::Debug + Clone + 'a,
-  CellKey : fmt::Debug + Clone,
-{
-
-  fn rows( &'a self ) -> impl IteratorTrait< Item = Row >
-  {
-    self.as_ref().fields().map( move | ( _k, e ) | e.into_owned() )
-  }
-
-}
-
-impl< 'a, T, RowKey, Row, CellKey, Cell > TableHeader< 'a, CellKey, CellKey >
-for AsTable< 'a, T, RowKey, Row, CellKey, Cell, CellKey >
-where
-  Self : TableRows< 'a, RowKey, Row, CellKey, Cell >,
-  Row : Clone + for< 'cell > Cells< 'cell, CellKey, Cell > + 'a,
-  CellKey : fmt::Debug + Clone,
-  Cell : fmt::Debug + Clone + 'a,
-  CellKey : fmt::Debug + Clone,
-{
-
-  fn header( &'a self ) -> Option< impl IteratorTrait< Item = ( CellKey, CellKey ) > >
-  {
-    let mut rows = self.rows();
-    let row = rows.next();
-    if let Some( row ) = row
-    {
-      Some
-      (
-        row
-        .cells()
-        .map( | ( key, _title ) | ( key.clone(), key ) )
-        .collect::< Vec< _ > >()
-        .into_iter()
-      )
-    }
-    else
-    {
-      None
-    }
-  }
-
-}
-
-impl< 'a, Row, CellKey, Cell > Cells< 'a, CellKey, Cell >
-for Row
-where
-  Row : Fields< 'a, CellKey, Cell > + 'a,
-  Cell : fmt::Debug + Clone + 'a,
-{
-
-  fn cells( &'a self ) -> impl IteratorTrait< Item = ( CellKey, Cell ) >
-  {
-    self.fields().map( move | ( key, cell ) | ( key, cell.into_owned() ) )
-  }
-
-}
-
-// ==
+//=
 
 /// Struct to hold options to print data as table.
 #[ derive( Debug, Default, Former ) ]
@@ -144,13 +14,13 @@ pub struct Styles
 }
 
 /// Struct for formatting tables.
-pub struct Formatter< 'a >
+pub struct Context< 'a >
 {
   buf : &'a mut dyn fmt::Write,
   styles : Styles,
 }
 
-impl< 'a > Formatter< 'a >
+impl< 'a > Context< 'a >
 {
   /// Just constructr.
   pub fn new( buf : &'a mut dyn fmt::Write, styles : Styles ) -> Self
@@ -159,12 +29,12 @@ impl< 'a > Formatter< 'a >
   }
 }
 
-impl fmt::Debug for Formatter< '_ >
+impl fmt::Debug for Context< '_ >
 {
   fn fmt( &self, f : &mut fmt::Formatter< '_ > ) -> fmt::Result
   {
     f
-    .debug_struct( "Formatter" )
+    .debug_struct( "Context" )
     .field( "buf", &"dyn fmt::Write" )
     .field( "styles", &self.styles )
     .finish()
@@ -189,7 +59,7 @@ where
   fn table_to_string( &'a self ) -> String
   {
     let mut output = String::new();
-    let mut formatter = Formatter
+    let mut formatter = Context
     {
       buf : &mut output,
       styles : Styles::default(),
@@ -208,7 +78,7 @@ where
 pub trait TableFormatter< 'b >
 {
   /// Formats the table and writes the result to the given formatter.
-  fn fmt< 'a >( &'b self, f : &mut Formatter< 'a > ) -> fmt::Result;
+  fn fmt< 'a >( &'b self, f : &mut Context< 'a > ) -> fmt::Result;
 }
 
 /// A trait for formatting tables.
@@ -223,7 +93,7 @@ where
   Cell : fmt::Debug + Clone + 'a,
   CellKey : fmt::Debug + Clone,
 {
-  fn fmt( &'a self, f : &mut Formatter< '_ > ) -> fmt::Result
+  fn fmt( &'a self, f : &mut Context< '_ > ) -> fmt::Result
   {
     let table_size = self.table_size();
     let mut col_widths : Vec< usize > = vec![ 0 ; table_size[ 1 ] ];
